api upgraded

+++ server/controllers/products/deleteProduct.js
@@ -8,6 +8,15 @@ const deleteProduct = async (req, res, next) => {
     const admin = await AdminModel.find({ uid: uid }).select(
       "-_id -role -totalearning -productpending -productdeliverd -delivertoadmin -order -ordercencle -username -email -uid  -__v -shopname"
     );
+
+    const findproduc = await Products.find({ name: Name, shopname: shopname })
+    if (!findproduc) {
+      res.status(400).json({
+        message: "product not found"
+      })
+      return
+    }
+
     await Products.findOneAndDelete({ name: Name, shopname: shopname });
     const products = admin[0].products;
     const filtered = products.filter((data) => data.name !== Name);
